post data to sql db using axios
As stated in our issue 59 (https://github.com/privacy-tech-lab/gpc-web-crawler/issues/59), it'd be great to be able to use an sql db to store data for our crawling purposes. I implemented one to try it out. The rest api to use it is in our repo. More details in our issue 59.